backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
djangorestframework-simplejwt>=5.3.1
django-cors-headers
dj-rest-auth
gunicorn
Pillow
cryptography>=45.0.5
paho-mqtt>=1.6.0